add losses and impurity

+++ java/src/main/java/loss/AbstractLoss.java
@@ -5,7 +5,7 @@ import java.util.List;
 public abstract class AbstractLoss implements Loss {
 
     @Override
-    public double batchLoss(List<Double> estimates, List<Double> groundTruths) {
+    public double batchLoss(List<Double> estimates, List<Integer> groundTruths) {
         assert estimates.size() == groundTruths.size();
         double result = 0;
         for (int i = 0; i < estimates.size(); i++) {
